+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Link casing to backfill.
 - Removed square brackets from _m MD_ in `chronostratigraphy` and `lithostratigraphy`.
 - Implemented a validation to check the user's role and the lock status of the borehole before performing any data manipulations on child tables in the .NET API.
+- Include completion in casing name for hydrogeological observations.
 
 ### Fixed